Project 007 - Crew AI Agents x RAG/requirements.txt
google-generativeai
python-dotenv
streamlit
langchain
langchain_community
langchain-google-genai
tavily-python
faiss-cpu
beautifulsoup4
pypdf